AgroData/requirements.txt
Flask==2.3.0
matplotlib==3.7.0
scipy==1.10.0
mysql-connector-python==9.0.0
pandas==2.2.3
numpy==1.26.4
Flask-Login==0.6.3
passlib[bcrypt]==1.7.4
